backend/requirements.txt
# Core Dependencies
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0

# Geospatial Processing
rasterio==1.3.9
geopandas==0.14.1
shapely==2.0.2
pyproj==3.6.1
fiona==1.9.5

# Image Processing
Pillow==10.1.0
opencv-python==4.8.1.78
scikit-image==0.22.0

# Machine Learning
tensorflow==2.15.0
numpy==1.24.3
scikit-learn==1.3.2

# Sentinel API (using sentinelsat for open access)
sentinelsat==1.2.1
requests==2.31.0

# Google Gemini API
google-generativeai==0.3.2

# Utilities
tqdm==4.66.1
pyyaml==6.0.1
